requirements.txt
amqp==2.5.1
analytics==0.6.5
authy==2.2.3
Babel==2.7.0
billiard==3.6.1.0
blinker==1.4
braintree==3.57.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.3.0
certifi==2017.7.27.1
cffi==1.13.2
chardet==3.0.4
Click==7.0
conda==4.3.16
convertdate==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect2==0.2.2
cycler==0.10.0
Cython==0.29.14
defusedxml==0.6.0
dill==0.3.1.1
Django==2.2.4
django-braces==1.13.0
django-dotenv==1.4.1
django-localflavor==2.2
django-parler==2.0
django-perms==2.1.0
django-rest-swagger==2.1.1
django-rosetta==0.9.3
django-suit==0.2.28
django-taggit==1.1.0
djangorestframework==3.10.3
fbprophet==0.5
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
flower==0.9.3
gunicorn==20.0.2
holidays==0.9.11
html5lib==1.0.1
idna==2.6
importlib-metadata==0.23
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
kombu==4.6.5
lunardate==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
more-itertools==7.2.0
ndg-httpsclient==0.5.1
nexmo==1.5.0
nltk==3.4.5
numpy==1.17.3
nydus==0.11.0
oauthlib==3.1.0
openapi-codec==1.3.2
pandas==0.25.1
pandas-datareader==0.8.1
passlib==1.7.2
phonenumbers==8.10.16
Pillow==6.2.1
plotly==4.2.1
polib==1.1.0
psycopg2==2.8.4
pusher==2.1.4
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.19
PyJWT==1.7.1
PyMeeus==0.3.6
pymongo==3.9.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
Pyphen==0.9.5
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-stdnum==1.12
python3-openid==3.1.0
pytz==2017.2
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rosetta==0.3
ruamel.yaml==0.16.5
scikit-learn==0.23.dev0
scipy==1.3.1
setuptools-git==1.2
simplejson==3.11.1
six==1.13.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
speaklater==1.3
SQLAlchemy==1.3.11
sqlparse==0.3.0
tinycss2==1.0.2
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
WeasyPrint==50
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
